Move Modifier Extension with Shimmer Effect to Core Module
